memefly-de/requirements.txt
pandas
pylint
autopep8
urllib3
jupyter
beautifulsoup4
requests
python-decouple
pymongo
dvc
scipy